Update MAC_SDK from 9.04 to 10.26
- Original download location:
  https://monkeysaudio.com/files/MAC_1026_SDK.zip
  SHA256:
  cff58b2ee049bfbde131b050a7fabd216b1618d86ada3dd956e542490ae64394
- Update the CUETools specific MACLibDll.cpp and MACLibDll.h based on
  the upstream modifications to MACDll.cpp and MACDll.h, respectively.
  * c_APEDecompress_GetData:
    Use `unsigned char * pBuffer` instead of `char * pBuffer`
    Remark: The equivalent of unsigned char in C# is byte.
  * Rename macros from MAC_ to APE_ according to upstream (10.18) [1].
    MACLibDll.h:
      MAC_COMPRESSION_LEVEL_NORMAL -> APE_COMPRESSION_LEVEL_NORMAL
    MACLibDll.cpp:
      MAC_FILE_VERSION_NUMBER -> APE_FILE_VERSION_NUMBER
      MAC_VERSION_STRING -> APE_VERSION_STRING
  * `Seek()` is used again in IO.h since 10.00. Furthermore it returns
    int instead of int64 since 10.01 [1].
- Update ThirdParty_MAC_SDK_CUETools.patch accordingly
- The fix of the exception when decoding very small APE files (316f907)
  has been submitted upstream and merged in the meantime (9.10).
- Resolves #261

[1] https://monkeysaudio.com/versionhistory.html

+++ CUETools.Codecs.MACLib/StreamIO.cs
@@ -76,7 +76,7 @@ namespace CUETools.Codecs.MACLib
             return m_stream.Length;
         }
 
-        long SeekRelativeCallback(void* id, long delta, int mode)
+        int SeekRelativeCallback(void* id, long delta, int mode)
         {
             m_stream.Seek(delta, (SeekOrigin)(mode));
             return 0;
